% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_cumulative_cases.R
\name{get_cumulative_cases}
\alias{get_cumulative_cases}
\title{Get cumulative case count as sum of specified case categories}
\usage{
get_cumulative_cases(x, case_type = c("scc", "sc", "cc", "scd", "sd",
  "cd", "ALL"))
}
\arguments{
\item{x}{data from ProMED and HealthMap.
Should only contain numeric columns called "sc", "cc", "sd", "cd".}

\item{case_type}{quoted character. should be one of scc, sc, cc, scd, sd, cd or ALL.}
}
\value{
numeric vector of cumulative case counts.
}
\description{
Data from ProMED and HealthMap contain counts for suspected
and confirmed cases, and suspected and confirmed deaths.
This function adds up the case counts in the specified categories
as the cumulative case count. The categories are:
}
\details{
scc: suspected and confirmed cases
sc: suspected cases
cc: confirmed cases
scd: suspected and confirmed deaths
sd: suspected deaths
cd: confirmed deaths
ALL: suspected and confirmed cases, and suspected and confirmed deaths

The cumulative case count for a date will be NA if ALL counts in the specified categories
on are NA for this date. Otherwise, it is the sum of the available values with NAs removed.
}
\author{
Sangeeta Bhatia
}
